Extend the rename implementation for species
Species (and prefixes/names) now allow a more general form or renaming,
which also provides a witness of the variables existence to the renaming
function.

For instance,

  species.rename_with : (A : species x)
                     -> (f : (a : name x) -> free_in a A -> name y)
                     -> species y

This allows us to safely _lower_ all variables (rather than raise them),
if we have a witness that no level 0 variables exist - as otherwise we
would have a contradiction.