Redesign equivalence to remove explicit symmetry
While this is convenient some times, it makes it quite hard to write
proofs of the form "A ~ B -> P A -> P B", as we also need a
"A ~ B -> P B -> P A" proof.

An alternative approach would be to define both forms of equivalency,
and define a bijection between the two, but this is sufficient for now.