Also require the affinity network is symmetric
This is specified in the original paper, but I entirely failed to state
this requirement in the type. This is needed in order to have T<a, b>
mean the same thing as T<b, a>.

This also redefines order to use a plain definition rather than an
inductive type. While this is rather ugly, I was having trouble getting
the equation compiler to work with the modifications.