Show transitions hold over equivalency
Well, sort of - there's one axiom in here (namely that if
`(rename A) [a]--> E`, then there is some a', E' such that
`[A] [a']--> E'`, and rename a' = a, rename E' = E.

I'm 99% sure this is true, but proving it right now has shown to be
incredibly annoying, and at this point I just want to be done with this
particular theorem. Sorry!