Convert prefix "context telescopes" to an actual datatype
Previously prefixes were defined as `context -> (context -> context) ->
Type`, where this function potentially modified the current context.
However, functions are rather annoying to do proofs on, and so we now
use an explicit "telescope" datatype.

One side effect of this is that we can now perform case splits on
transitions within the enumerate proof - previously it was this function
which was blocking this from occuring.