Push our generalisation of equivalence through
Oh dear, this is going to cause complications. The main problem here is
that we need a lemma which says:

   For all transitions A [a]--> E, for any equivalent species B there is
   a transition B [a]--> E', such that E and E' are equivalent.

I had assumed that we could get away with just having "equivalence
relation R implies structural congruence" and we could just use our
proof of this for congruence.

However, this is not the case, as need to map the proof that "E is
congruent to E'" back to our normal equivalence relation.

It's /really/ not clear how we could even prove this original lemma for
n-equivalence - we'd have to do some induction on the structure of both
species, and this does not feel like a bundle of fun.