Redefine transitions a little
This effectively tries to remove the complexity in constructor indices,
in an attempt to simplify proofs.

 - Remove the implicit casts. While conceptually neat, they get in the
   way of actually showing things.

 - Remove production.map, and split up rules using it into "production"
   and "species" variants. This does lead to some duplication, but it's
   not the end of the world.

Note, this comments out much of the proof for "transitions.equivalence"
(though it is now slightly shorter!). I'll get to it at some point, but
nothing interesting there needs to be shown, it's just redoing some of
the proof.