Allow enumerating transitions from choices
We've been able to non-computably do this for ages. This finally drops
the requirement of a decidable_eq instance, which allows us to actually
execute it.

These functions are rather slow to type check (in the realms of 1-4s, so
we split them up to make it a tiny bit more reasonable).